+++ backend/src/main/kotlin/com/pk/zbtz/zbtzbackend/utils/execution_timer/ExecutionTimer.kt
@@ -1,4 +1,4 @@
-package com.pk.zbtz.zbtzbackend.utils
+package com.pk.zbtz.zbtzbackend.utils.execution_timer
 
 interface ExecutionTimer {
     fun <T> measure(block: () -> T): ElapsedTimeResult<T>
